% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQthemes.R
\docType{data}
\name{aq_color_good_bad}
\alias{aq_color_good_bad}
\title{This is a binary Good/Bad color palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
aq_color_good_bad
}
\description{
This is a binary Good/Bad color palette
}
\keyword{datasets}
